requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
csvImporter==0.1.3.2
Django==2.1.7
django-adaptors==0.2.5
django-cors-headers==2.4.1
django-filter==2.1.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-request-logging==0.6.7
djangorestframework==3.9.1
drf-writable-nested==0.5.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.1.1
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
PyYAML==3.13
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1